adding database connection

+++ src/main/java/motorola/akademia/shop/apiControllers/OrderApiController.java
@@ -1,8 +1,8 @@
 package motorola.akademia.shop.apiControllers;
 
 import lombok.AllArgsConstructor;
-import motorola.akademia.shop.repository.Cart;
-import motorola.akademia.shop.repository.Order;
+import motorola.akademia.shop.entities.Cart;
+import motorola.akademia.shop.entities.Order;
 import motorola.akademia.shop.services.OrderListService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -29,7 +29,7 @@ public class OrderApiController {
     }
 
     @GetMapping("/getOrderById/{id}")
-    public ResponseEntity<Order> getOrdersByUser(@PathVariable int id){
+    public ResponseEntity<Order> getOrdersByUser(@PathVariable Long id){
         return ResponseEntity.ok(orderListService.getOrdersById(id));
     }
